Move opengauss related classes from scaling-opengauss to cdc-opengauss
