+++ shardingsphere-kernel/shardingsphere-schedule/shardingsphere-cdc/shardingsphere-cdc-dialect/shardingsphere-cdc-opengauss/pom.xml
@@ -31,8 +31,22 @@
     <dependencies>
         <dependency>
             <groupId>org.apache.shardingsphere</groupId>
-            <artifactId>shardingsphere-cdc-core</artifactId>
-            <version>${project.parent.version}</version>
+            <artifactId>shardingsphere-cdc-postgresql</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>com.google.code.gson</groupId>
+            <artifactId>gson</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>com.zaxxer</groupId>
+            <artifactId>HikariCP</artifactId>
+            <scope>compile</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.opengauss</groupId>
+            <artifactId>opengauss-jdbc</artifactId>
+            <scope>provided</scope>
         </dependency>
     </dependencies>
 </project>
