+++ shardingsphere-kernel/shardingsphere-schedule/shardingsphere-cdc/shardingsphere-cdc-dialect/shardingsphere-cdc-opengauss/src/test/java/org/apache/shardingsphere/cdc/opengauss/wal/decode/MppdbDecodingPluginTest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.scaling.opengauss.wal.decode;
+package org.apache.shardingsphere.cdc.opengauss.wal.decode;
 
 import com.google.gson.Gson;
 import lombok.SneakyThrows;
